Add files via upload
This is the first beta of the speed density file for 3B ECU's. This does require a patch wire to feed the MAP signal from the boost side to the motor chip processor. At the current time that part is not defined in this file as an ADC input has not been chosen for a MAP sensor input as of yet.